requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
jinja2>=3.1.2
python-jose>=3.3.0
python-dotenv>=1.0.0
boto3>=1.26.100
httpx>=0.24.0
email-validator>=2.0.0
pydantic>=1.10.7
python-multipart>=0.0.6
